requirements.txt
openai
chonkie[all]
uvicorn
fastapi
pydantic
transformers
python-dotenv
cachetools